Check for error from grant handler
If the refresh token was created with an access token that is currently in a state of ACTIVE, but both the access token and refresh token are aged beyond their validity periods, the refresh grant handler will set the error info on the response and return it with a null access token.  If this check isn't performed and the rest of the method runs it will get a null pointer exception trying to build the id token.