IDENTITY-5182: Improve Caching configuration in OAuth
This fix is to deprecate <EnableOAuthCache> configuration in identity.xml being used to decide the caching beheviour in OAuth.
Since we have provided a separate mechanism to control each cache type in OAuth using <CacheConfig> element in identity.xml we do not need
to consider <EnableOAuthCache> property to check whether any cache is on/off.